requirements.txt
click==8.3.1
markdown-it-py==4.0.0
mdurl==0.1.2
pillow==12.1.0
Pygments==2.19.2
rich==14.3.1
shellingham==1.5.4
typer==0.21.1
typing_extensions==4.15.0
